task: split URI & URI.Template into separate files
